Changes to Avizo angles for consistent typing and core XTek routines to properly set up geometry. Reader/Filter/Reconstruction modules need further testing but no longer seem to crash.

+++ Avizo/src/CCPiReconstruction/XtekReader.h
@@ -48,7 +48,7 @@ private:
 	//_ctdata.txt parameters
 	int framesPerProjection;
 	double exposureTime;
-	double *angles;
+	float *angles;
 	double *angleTime;
 	//Image data
 	float* imageData;
@@ -74,7 +74,7 @@ public:
 	int getImageHeight(){return detectorPixels[1];}
 	int getFramesPerProjection(){return framesPerProjection;}
 	double getExposureTime(){return exposureTime;}
-	double* getAngles(){return angles;}
+	float* getAngles(){return angles;}
 	double* getVoxelSize(){return voxelSize;}
 	double* getOffset(){return offset;}
 	double getSourceToObject(){return srcToObject;}
